Rename filtration to filters
changed direct paths to relative paths
