+++ eform-client/src/app/plugins/modules/items-planning-pn/components/plannings/plannings-page/plannings-table/plannings-table.component.ts
@@ -9,7 +9,7 @@ import {
 import { PlanningModel } from '../../../../models/plannings';
 import { Paged, TableHeaderElementModel } from 'src/app/common/models';
 import { ItemsPlanningPnClaims } from '../../../../enums';
-import { PlanningsStateService } from '../../store/';
+import { PlanningsStateService } from '../../store';
 import { AuthStateService } from 'src/app/common/store';
 
 @Component({
