+++ eform-client/src/app/plugins/modules/items-planning-pn/store-providers.config.ts
@@ -1,5 +1,5 @@
-import { planningsPersistProvider } from './components/plannings/store/plannings.store';
-import { pairingPersistProvider } from './components/pairing/store/pairing-store';
+import { planningsPersistProvider } from './components/plannings/store';
+import { pairingPersistProvider } from './components/pairing/store';
 
 export const planningsStoreProviders = [
   planningsPersistProvider,
